sl-API-backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
bcrypt==4.0.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.121.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
MarkupSafe==3.0.3
marshmallow==3.26.1
motor==3.7.1
packaging==25.0
passlib==1.7.4
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.10.1
pymongo==4.15.3
python-decouple==3.8
python-dotenv==1.2.1
python-http-client==3.3.7
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
sendgrid==6.12.5
sniffio==1.3.1
starlette==0.49.3
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
Werkzeug==3.1.3
